register woocommerce checkout and cart blocks
